Add a new service to handle the change of a receiver module
Changing a receiver module triggers a (persistent) notification offering
the user a new calibration. This offer cannot be ignored, only delayed.
it will show up again on next boot or after one day if the user asks for a
reminder. Completing the calibration will remove the notification for good.

When clicked, the notification will start a new activity displaying the two
possible actions: setting a reminder or executing the calibration. If the
user chooses so, the regular calibration activity is started.

FPIIM-1575

Change-Id: Ia12dcf87eca06f115212e01368b457465fdd01fa