Remove unused STOP_APP_SWITCHES permission
The STOP_APP_SWITCHES permission was introduced in commit 707feff. The
permission allows the app to bypass Android's five-second delay for
starting an activity after the app has been put in the background by the
user. The current implementation of the app does not require such a
bypass, so remove the related permission.

Issue: FP2P-420
Change-Id: I9fc88a130e657dc3912ac0dc5e6a973a0c4769c2